Move mmap type module to common
