+++ lib/segment/src/common/mmap_type.rs
